requirements.txt
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
boto3>=1.34.0
google-adk>=0.1.0
google-genai>=0.1.0